game-agent/requirements.txt
chainlit
python-dotenv
openai
agents 